Moved getRoutes() to RoutingWP
