+++ layers/Engine/packages/root/src/Routing/AbstractRoutingManager.php
@@ -13,44 +13,6 @@ abstract class AbstractRoutingManager implements RoutingManagerInterface
     use BasicServiceTrait;
 
     /**
-     * @var string[]|null
-     */
-    private ?array $routes = null;
-
-    /**
-     * @return string[]
-     */
-    public function getRoutes(): array
-    {
-        if ($this->routes === null) {
-            $this->routes = array_filter(
-                (array) App::applyFilters(
-                    RouteHookNames::ROUTES,
-                    []
-                )
-            );
-
-            // // If there are partial endpoints, generate all the combinations of route + partial endpoint
-            // // For instance, route = "posts", endpoint = "/api/rest", combined route = "posts/api/rest"
-            // if ($partialEndpoints = array_filter(
-            //     (array) \PoP\Root\App::applyFilters(
-            //         'route-endpoints',
-            //         []
-            //     )
-            // )) {
-            //     // Attach the endpoints to each of the routes
-            //     $routes = $this->routes;
-            //     foreach ($routes as $route) {
-            //         foreach ($partialEndpoints as $endpoint) {
-            //             $this->routes[] = $route . '/' . trim($endpoint, '/');
-            //         }
-            //     }
-            // }
-        }
-        return $this->routes;
-    }
-
-    /**
      * By default, everything is a generic route
      */
     public function getCurrentNature(): string
