+++ layers/Engine/packages/root/src/Routing/RoutingManagerInterface.php
@@ -6,10 +6,6 @@ namespace PoP\Root\Routing;
 
 interface RoutingManagerInterface
 {
-    /**
-     * @return string[]
-     */
-    public function getRoutes(): array;
     public function getCurrentNature(): string;
     public function getCurrentRoute(): string;
 }
